requirements.txt
Django==1.10.1
django-picklefield
celery
wheel==0.29.0  # for build into wheel